+++ js/main.js
@@ -1,4 +1,2 @@
-// импорт всех файлов
-import './data.js';
-import './util.js';
-import './functions.js';
+import {getPhotoArray} from './post-maker.js';
+console.log(getPhotoArray());
